homework_psql/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.0
doxypypy==0.8.8.6
Flask==1.0.2
Flask-WTF==0.14.2
idna==2.6
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
nbformat==4.4.0
numpy==1.16.1
pandas==0.24.1
Pillow==5.4.1
plotly==2.5.1
psycopg2-binary==2.8.4
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.4
PyYAML==3.13
requests==2.18.4
six==1.11.0
SQLAlchemy==1.3.10
traitlets==4.3.2
urllib3==1.22
Werkzeug==0.14.1
wordcloud==1.5.0
WTForms==2.2.1
gunicorn
